Perform snapshot tests
