+++ package.json
@@ -8,7 +8,8 @@
     "@testing-library/user-event": "^7.2.1",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
-    "react-scripts": "3.4.1",
+    "react-scripts": "^3.4.3",
+    "react-test-renderer": "^16.13.1",
     "styled-components": "^5.1.1"
   },
   "scripts": {
